Create service/delete-ai-assistant.ts
